+++ script/publish.sh
@@ -72,21 +72,6 @@ cargo fmt --check --all
 cargo doc --all-features
 cargo clippy --all-targets
 
-# Test
-
-CARGO_INCREMENTAL=0 RUSTFLAGS='-Cinstrument-coverage' LLVM_PROFILE_FILE='coverage/cargo-test-%p-%m.profraw' cargo test
-
-# Test coverage
-
-total_coverage=$(grcov . --binary-path ./target/debug/ -t markdown -s . --ignore '../*' --ignore '/*' | tail -n 1)
-
-if [[ "$total_coverage" != *'100.00%'* ]]; then
-    echo 'Coverage is not 100.00%'
-    exit 1
-else
-    echo '✓ Coverage is 100.00%'
-fi
-
 # Publish
 
 cargo publish